+++ packages/story-group-composer/components/Carousel/CardItem/styles.jsx
